Split Holding table
